+++ src/main/java/de/cowtipper/cowlection/listener/skyblock/DungeonsListener.java
@@ -418,8 +418,9 @@ public class DungeonsListener {
                 }
             }
         }
-        if (partyType != DataHelper.PartyType.CURRENT
-                || (/*partyType == PartyType.CURRENT &&*/ Minecraft.getSystemTime() % 1000 < 600)) {
+        if (MooConfig.dungPartyFinderOverlayDrawBackground &&
+                (partyType != DataHelper.PartyType.CURRENT
+                        || (/*partyType == PartyType.CURRENT &&*/ Minecraft.getSystemTime() % 1000 < 600))) {
             GlStateManager.pushMatrix();
             GlStateManager.translate(0, 0, partyType.getZIndex());
             Gui.drawRect(x, y, x + 16, y + 16, partyType.getColor());