+++ CHANGELOG.md
@@ -36,6 +36,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Dungeon Party Finder: Parties with specific classes can now *always* be marked as 'unideal' (additionally to the already existing option to mark a party when 2+ members use the same specific class)
 - Dungeon Performance Overlay: added an alternative text border option
 - "Copy inventories to clipboard"-feature now automatically decodes base64 data (e.g. skin details) and unix timestamps
+- Dungeon Party Finder: colored overlay is now also disable-able via config
 
 ### Fixed
 - Fixed issue with 'no dung class selected'